+++ ErsatzTV.Application/Jellyfin/Commands/CallJellyfinLibraryScannerHandler.cs
@@ -64,6 +64,11 @@ public class CallJellyfinLibraryScannerHandler : CallLibraryScannerHandler<ISync
         {
             arguments.Add("--force");
         }
+        
+        if (request.DeepScan)
+        {
+            arguments.Add("--deep");
+        }
 
         return await base.PerformScan(scanner, arguments, cancellationToken);
     }